requirements.txt
wrf-python
netCDF4
cartopy
pandas
Pillow
matplotlib
xarray
scipy
xesmf
paramiko
python-dotenv